:recycle: refactor components to module
basic refactor and split codes
